Add optional component_type_id and component_id fields to dispatch table
These allow the user to specify the type(s) of component that should be
used to supply the dispatch, or alternatively the exact components, via
their specific IDs.  To make sure they can be used in the unique index,
they may not be `NULL`, but default to empty if not explicitly set.

Both of these fields are arrays whose entries are assumed to be a cross
reference to some external database.  If both fields are empty, then it
is assumed that the microgrid has free rein in deciding how to implement
the requested dispatch.  Similarly, if `component_type_id` is set while
`component_id` is not, then the microgrid has free rein in deciding the
exact components of those type(s) that will be used.  If `component_id`
is set, then the microgrid can choose between the different components,
but only those.

In short, these fields act as filters on the choices the microgrid can
make; if empty, there is no such filter.  To avoid inconsistencies, it
is recommended (required?) that at most one of these fields should be
non-empty.