Remove DispatchComponent
This message type was made as a result of misunderstood requirements. A
component does not necessarily have a type associated with it, and dispatch
does not care about this relationship. One "dispatch" should be able to be
applied to a specific set of components, or a specific set of component
types (e.g. charge all batteries)