Make sure to include .pyi files in the distribution
Since .pyi files are generated, we need to make sure they are part of
the distribution package (the tar.gz). This is done via the MANIFEST.in
file.

Once we do this, there is no need to specify them also in the
`tool.setuptools.package-data` configuration in `pyproject.toml` too,
as they will be included by default (same with the `py.typed` file).

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>