Remove unnecessary inclusion from MANIFEST
The MANIFEST.in file is used to create the distribution package, but the
*.pyi files should be generated *from* the distribution package, so they
are not present yet when the distribution package is built.

They should be included in the wheel though, so we need to make sure
they are listed as package data because setuptools doesn't consider them
automatically as sources for some reason.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>