Use `github-check` reporter for protolint
Using `github-pr-review` requires elevated write permissions to the
repository, just to write a review, which can compromise the security
of the repository, by using the `pull_request_target` event which gives
the GitHub token write access, or using a custom token which complicates
the setup.

To simplify things we switch back to `github-check` reporter, which is
the repo-config default and doesn't need any write permissions, and
failed checks will still be shown in the diff view.

We also pin the protolint version to `0.52.0` to avoid any surprises
with future versions, also as repo-config does.

https://github.com/reviewdog/reviewdog?tab=readme-ov-file#reporter-github-checks--reportergithub-check

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>